chore: split main & event_handler
